Remove thread local reader for stored fields
This replaces the thread local storage for stored fields with an
"accessor" object you fetch once per thread and let go out of scope when
you are done with it. This makes it much easier to reason about when the
state used to read the documents can be GCed: when the accessor goes out
of scope. With this we won't have to feel guilty keeping a decompressed
copy of the shared dictionary.

Before:
```
Document d1 = ir.document(1);
Document d2 = ir.document(1);
```

After:
```
StoredFields storedFields = ir.storedFields();
Document d1 = storedFields.document(1);
Document d2 = storedFields.document(1);
```